requirements.txt
langchain
langchain-community
langchain-openai
langchain-google-genai
langgraph
sqlalchemy

python-dotenv
flask
flask-cors
